Move CallHandler impl into call2
