+++ crates/call2/Cargo.toml
@@ -31,7 +31,7 @@ media = { path = "../media" }
 project = { package = "project2", path = "../project2" }
 settings = { package = "settings2", path = "../settings2" }
 util = { path = "../util" }
-
+workspace = {package = "workspace2", path = "../workspace2"}
 anyhow.workspace = true
 async-broadcast = "0.4"
 futures.workspace = true
